Clean up select reading from pipes
It doesn't make sense to use a while-loop or for-loop to continually try and
read from pipes that may or may not be written to multiple times over the
course of the process.

Instead, wait for the child process to finish, and then check if the pipes are
ready to read. If they are, copy results into a buffer.

A timeout on the processes themselves can be implemented later with the alarm
system call.